% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Fatality Analysis Reporting System (FARS) Data By State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer of 4 indicating the state number. Valid values
are 1 to 51, except 3}

\item{year}{A 4 digit integer or string indicating the year}
}
\value{
This function does not have any return value, but it will plot it will
   plot a heat map of the accident for the given state and year.
}
\description{
This is a function that takes an integer representing \code{state.num}
and 4 digit \code{year} as input and FARS data into a data frames using
\code{make_filename} and \code{fars_read}.
The FARS data is filtered by \code{state.num}, and if the filter criteria
does not exist in the dataset, the function will throw an error for invalid state.
The filtered data will be plotted as a heat map where area with more accidents
will be shown in darker spots. The function will give a message "no accident to plot"
if there is no accident to report for the state and year.
}
\note{
Though the function anticipates giving a message if no accident is reported
for a state in a given year, there is a very good chance that this portion of the
code will not be trigerred due to a check to ensure the \code{state.num} filter
parameter is in the data set.
}
\examples{
   \dontrun{fars_map_state(1,2015)}

}
