% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multi-year Fatality Analysis Reporting System (FARS) Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An vector of 4 digit integer or string indicating the year}
}
\value{
This function returns a list of data frame with FARS data for
   the years.
}
\description{
This is a function that takes a vector of 4 digit, called \code{years},
as input and read FARS data into a list of data frames, one for each year.
Each of the year input is converted to FARS filename via \code{make_filename}.
Each data frame only contains the year and month. If FARS data file
is not available any of the given year, the list entry for that year
will be NULL
}
\examples{
   \dontrun{fars_read_years(c(2014,2015))}
   \dontrun{fars_read_years((2013:2015))}

}
