% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize multi-year Fatality Analysis Reporting System (FARS) Data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An vector of 4 digit integer or string indicating the year}
}
\value{
This function returns a data frame with FARS data for summarized
   by years and months.
}
\description{
This is a function that takes a vector of 4 digit, called \code{years},
as input, read FARS data into a list of data frames, one for each year
via \code{fars_read_years}. These multi-year FARS data will be summarized
by counting the number of rows that exist for each year and month.
The summarized data frame will have months as rows, and years as columns, while
the row counts as the intersection of rows and columns.
}
\examples{
   \dontrun{fars_summarize_years(c(2014,2015))}
   \dontrun{fars_summarize_years((2013:2015))}

}
