% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct FARS filename from year input parameter}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A 4 digit integer or string indicating the year}
}
\value{
This function returns a FARS filename corresponding to the
   year.
}
\description{
This is a function that takes 4 digit \code{year} as input and
construct the FARS filename for the year.
}
\examples{
   \dontrun{make_filename(2015)}
   \dontrun{make_filename("2015")}

}
